+++ certCheck.py
@@ -1,21 +1,21 @@
 # Program:        Certificate Checker
 # Author:         Nolan Rumble
-# Date:           2022/06/16
-# Version:        0.18
+# Date:           2022/06/26
+# Version:        0.19
 
 import argparse
 import datetime
 import sys
 import json
 
-from systemInfo import systemInfo, systemData
+from systemInfo import systemInfo
 from certificate import certificateModule
 from data import certData
 from data import sendDataMongoDB
 from data import emailTemplateBuilder
 from data import sendDataEmail
 
-scriptVersion = "0.18"
+scriptVersion = "0.19"
 
 # Global Variables
 args = None
@@ -30,8 +30,8 @@ def parseArguments():
     parser = argparse.ArgumentParser(description='Certificate Checker v' + scriptVersion)
 
     # Optional arguments
-    parser.add_argument('--hostname', default="google.com",
-                        help='Hostname to get certificate from. Defaults to google.com')
+    parser.add_argument('--hostname', default='',
+                        help='Hostname to get certificate from')
 
     parser.add_argument('--displayCertificate', action='store_true',
                         help='Display certificate info')
@@ -58,33 +58,50 @@ def parseArguments():
                         help='Send an email with the results. SMTP connection details stored in mail.cfg')
 
     parser.add_argument('--setTag', default='',
-                        help='Set the tag for the query results. Creates tag.cfg file with tag. Use commas to separate multiple tags.')
+                        help='Set the tag for the query results. Use commas to separate multiple tags.')
 
-    parser.add_argument('--deleteTag', action='store_true',
-                        help='Delete the tag file - tag.cfg')
+    parser.add_argument('--delTag', action='store_true',
+                        help='Removes the tags from the configuration file.')
 
     parser.add_argument('--getTag', action='store_true',
-                        help='Get the tag from tag.cfg file')
+                        help='Get tags from the configuration file.')
 
-    parser.add_argument('--renewUuid', action='store_true',
-                        help='Renew the UUID value.')
+    parser.add_argument('--renewDeviceId', action='store_true',
+                        help='Renew the device UUID value.')
 
-    parser.add_argument('--getUuid', action='store_true',
-                        help='Get the UUID value from uuid.cfg file.')
+    parser.add_argument('--getDeviceId', action='store_true',
+                        help='Get the device UUID value from configuration file.')
 
-    parser.add_argument('--deleteUuid', action='store_true',
-                        help='Remove the UUID value. Caution: when script runs again a new UUID will be generated.')
+    parser.add_argument('--deleteDeviceId', action='store_true',
+                        help='Remove the device UUID value. When script runs again a new UUID will be generated.')
+
+    parser.add_argument('--setTenantId', default='',
+                        help='Sets the tenant ID for the script.')
+
+    parser.add_argument('--getTenantId', action='store_true',
+                        help='Gets the tenant ID from configuration file.')
+
+    parser.add_argument('--delTenantId', action='store_true',
+                        help='Deletes the tenant ID from the configuration file.')
+
+    parser.add_argument('--createBlankConfiguration', action='store_true',
+                        help='Creates a blank configuration file template - myConfig.json. Overwrites any existing configuration')
 
     global args
 
     args = parser.parse_args()
 
 
-def defineInfoArguments(o_systemData, o_systemInfo):
-    """This validates the arguments used for tag and uuid definitions."""
+def defineInfoArguments(o_systemInfo):
+    """This validates the arguments used for tag and device Id, and tenant Id definitions."""
+    # If createBlankConfiguration is set, create a blank configuration file template.
+    if args.createBlankConfiguration:
+        o_systemInfo.createBlankConfigurationFile()
+        sys.exit(0)
+
     # If setTag argument is set, create the new Tag.
     if args.setTag:
-        o_systemData.setTag(args.setTag)
+        o_systemInfo.setTag(args.setTag)
         print('New tag set.')
         sys.exit(0)
 
@@ -93,25 +110,40 @@ def defineInfoArguments(o_systemData, o_systemInfo):
         print(o_systemInfo.getTag())
         sys.exit(0)
 
-    # If deleteTag is set, it will delete the tag.cfg file.
-    if args.deleteTag:
-        o_systemData.deleteTag()
+    # If delTag is set, it will delete the tag(s) from the configuration file.
+    if args.delTag:
+        o_systemInfo.deleteTag()
         sys.exit(0)
 
-    # If getUuid is set, it grab the value in uuid.cfg
-    if args.getUuid:
-        print(o_systemInfo.getUuid())
+    # If getDeviceId is set, retrieve it from the configuration file.
+    if args.getDeviceId:
+        print(o_systemInfo.getDeviceId())
         sys.exit(0)
 
-    # If deleteUuid is set, the uuid.cfg file will be deleted.
-    if args.deleteUuid:
-        o_systemData.deleteUuid()
+    # If deleteDeviceId is set, remove it from the configuration file.
+    if args.deleteDeviceId:
+        o_systemInfo.deleteDeviceId()
         sys.exit(0)
 
-    # If renewUuid is set, first delete uuid.cfg file, then generate a new uuid.
-    if args.renewUuid:
-        o_systemData.deleteUuid()
-        o_systemData.createUuidIfNotExist()
+    # If renewDeviceId is set, delete it from the configuration file.
+    if args.renewDeviceId:
+        o_systemInfo.deleteDeviceId()
+        o_systemInfo.generateDeviceId()
+        sys.exit(0)
+
+    # If setTenantId is set, add it to the configuration file.
+    if args.setTenantId:
+        o_systemInfo.setTenantId(args.setTenantId)
+        sys.exit(0)
+
+    # If getTenantId is set, retrieve the tenant Id from the configuration file.
+    if args.getTenantId:
+        print(o_systemInfo.getTenantId())
+        sys.exit(0)
+    
+    # If delTenantId is set, remove the tenant Id from the configuration file.
+    if args.delTenantId:
+        o_systemInfo.deleteTenantId()
         sys.exit(0)
 
 
@@ -121,26 +153,23 @@ def gatherData(certResults):
     help with measuring results across multiple executions.
 
     Data that is included is:
-    * deviceUuid         - a unique device identifier.
-    * deviceTag          - a tag for the device to help with aggregating data across.
-                           multiple endpoints (for example all production, development, qa devices).
-    * hostName           - the hostname of the device where script is executing.
-    * scriptUTCStartTime - script start time (UTC format).
-    * scriptUTCEndTime   - script end time (UTC format).
-    * queryResults       - The results of all queries that were performed against the nameservers.
+    * myTenantId           - a unique tenant identifier.
+    * myDeviceId           - a unique device identifier.
+    * myDeviceTag          - a tag for the device to help with aggregating data across.
+                             multiple endpoints (for example all production, development, qa devices).
+    * hostName             - the hostname of the device where script is executing.
+    * scriptUTCStartTime   - script start time (UTC format).
+    * scriptUTCEndTime     - script end time (UTC format).
+    * queryResults         - The results of all queries that were performed against the nameservers.
     """
     myInfo = systemInfo.systemInfo()
 
-    if myInfo.uuid == "":
-        n = systemData.systemData()
-        n.createUuidIfNotExist()
-        myInfo.uuid = myInfo.getUuid()
-
     myData = {
-        "deviceUuid": myInfo.uuid,
-        "deviceTag": myInfo.deviceTag,
+        "tenantId": myInfo.myConfigJson["myTenantId"],
+        "deviceId": myInfo.myConfigJson["myDeviceId"],
+        "deviceTag": myInfo.myConfigJson["myTags"],
         "clientHostName": myInfo.hostname,
-        "dataFormatVersion": 6,
+        "dataFormatVersion": 7,
         "certResults": certResults
     }
 
@@ -304,11 +333,10 @@ if __name__ == "__main__":
     parseArguments()
 
     # Set initial objects for systemData and systemInfo
-    o_mySystemData = systemData.systemData()
     o_myInfo = systemInfo.systemInfo()
 
     # Gather system data and info
-    defineInfoArguments(o_mySystemData, o_myInfo)
+    defineInfoArguments(o_myInfo)
 
     # initialize jsonCertificateInfo
     jsonCertificateInfo = {}